// src/telegram/interfaces/session-state.enum.ts
export enum SessionState {
  NONE = 'NONE',
  AWAITING_EMAIL = 'AWAITING_EMAIL',
  AWAITING_OTP = 'AWAITING_OTP',
  AUTHENTICATED = 'AUTHENTICATED',
  AWAITING_RECIPIENT_EMAIL = 'AWAITING_RECIPIENT_EMAIL',
  AWAITING_AMOUNT = 'AWAITING_AMOUNT',
  AWAITING_CURRENCY = 'AWAITING_CURRENCY',
  AWAITING_MESSAGE = 'AWAITING_MESSAGE',
  AWAITING_WALLET_ADDRESS = 'AWAITING_WALLET_ADDRESS',
  AWAITING_NETWORK = 'AWAITING_NETWORK',
  AWAITING_BANK_SELECTION = 'AWAITING_BANK_SELECTION',
  AWAITING_CONFIRMATION = 'AWAITING_CONFIRMATION',
}